+++ examples/dane/src/mastra/agents/index.ts
@@ -9,7 +9,7 @@ import { imageTool } from '../tools/image.js';
 import { readPDF } from '../tools/pdf.js';
 import { activeDistTag, pnpmBuild, pnpmChangesetPublish, pnpmChangesetStatus } from '../tools/pnpm.js';
 
-import { getBaseModelConfig } from './model';
+import { getBaseModelConfig } from './model.js';
 
 export const daneCommitMessage = new Agent({
   name: 'DaneCommitMessage',
@@ -32,11 +32,27 @@ export const daneIssueLabeler = new Agent({
   model: getBaseModelConfig(),
 });
 
+const packages_llm_text = `
+  @mastra/core is located in the "packages/core" directory.
+  @mastra/deployer is located in the "packages/deployer" directory.
+  mastra is located in the "packages/cli" directory.
+  @mastra/engine is located in the "packages/engine" directory.
+  @mastra/evals is located in the "packages/evals" directory.
+  @mastra/rag is located in the "packages/rag" directory.
+  @mastra/tts is located in the "packages/tts" directory.
+  @mastra/memory is located in the "packages/memory" directory.
+  @mastra/mcp is located in the "packages/mcp" directory.
+  @mastra/deployer-{name} is located in the "deployers/{name}" directory.
+  dane or @mastra/dane is located in the "exampeles/dane" directory.
+`;
+
 export const danePackagePublisher = new Agent({
   name: 'DanePackagePublisher',
   instructions: `
     You are Dane, the ultimate node module publisher.
     You help engineers publish their pnpm changesets.
+
+    ${packages_llm_text}
     `,
   model: getBaseModelConfig(),
   tools: {