+++ packages/deployer/src/server/handlers/agents.ts
@@ -2,6 +2,7 @@ import { Context } from 'hono';
 import { stringify } from 'superjson';
 import zodToJsonSchema from 'zod-to-json-schema';
 
+import { readFile } from 'fs/promises';
 import { HTTPException } from 'hono/http-exception';
 
 import { handleError } from './error';
@@ -66,6 +67,42 @@ export async function getAgentByIdHandler(c: Context) {
   }
 }
 
+export async function getEvalsByAgentIdHandler(c: Context) {
+  try {
+    const mastra = c.get('mastra');
+    const agentId = c.req.param('agentId');
+    const agent = mastra.getAgent(agentId);
+    const evals = await readFile('./evals.json', 'utf-8');
+    const parsedEvals = evals
+      .split('\n')
+      .map(line => line && JSON.parse(line))
+      .filter((line: any) => line?.meta?.agentName === agent.name);
+    return c.json({
+      ...agent,
+      evals: parsedEvals,
+    });
+  } catch (error) {
+    return handleError(error, 'Error getting evals');
+  }
+}
+
+export function getLiveEvalsByAgentIdHandler(evalStore: any) {
+  return async (c: Context) => {
+    try {
+      const mastra = c.get('mastra');
+      const agentId = c.req.param('agentId');
+      const agent = mastra.getAgent(agentId);
+      const parsedEvals = evalStore.filter((line: any) => line?.meta?.agentName === agent.name);
+      return c.json({
+        ...agent,
+        evals: parsedEvals,
+      });
+    } catch (error) {
+      return handleError(error, 'Error getting evals');
+    }
+  };
+}
+
 export async function generateHandler(c: Context) {
   try {
     const mastra = c.get('mastra');
@@ -84,6 +121,7 @@ export async function generateHandler(c: Context) {
     }
 
     const result = await agent.generate(messages, { threadId, resourceid, output });
+
     return c.json(result);
   } catch (error) {
     return handleError(error, 'Error generating from agent');
@@ -101,6 +139,7 @@ export async function streamGenerateHandler(c: Context) {
     }
 
     const { messages, threadId, resourceid, output } = await c.req.json();
+
     validateBody({ messages });
 
     if (!Array.isArray(messages)) {
@@ -108,6 +147,7 @@ export async function streamGenerateHandler(c: Context) {
     }
 
     const streamResult = await agent.stream(messages, { threadId, resourceid, output });
+
     return new Response(streamResult.toDataStream(), {
       headers: {
         'Content-Type': 'text/x-unknown',
@@ -116,6 +156,7 @@ export async function streamGenerateHandler(c: Context) {
       },
     });
   } catch (error) {
+    console.log(error);
     return handleError(error, 'Error streaming from agent');
   }
 }