+++ deployers/cloudflare/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @mastra/deployer-cloudflare
 
+## 0.0.1-alpha.8
+
+### Patch Changes
+
+- 1944807: Unified logger and major step in better logs
+- c7abf8e: Optional CF worker tagging
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/deployer@0.0.1-alpha.7
+  - @mastra/core@0.1.27-alpha.69
+
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [291fe57]
+- Updated dependencies [1a41fbf]
+  - @mastra/deployer@0.0.1-alpha.6
+
+## 0.0.1-alpha.6
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/deployer@0.0.1-alpha.5
+
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [7babd5c]
+  - @mastra/deployer@0.0.1-alpha.4
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- 026ca5d: Optional dispatcher namespace configuration for cloudflare deployments
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- a291824: Deployer fixes
+- Updated dependencies [c8ff2f5]
+- Updated dependencies [a291824]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/deployer@0.0.1-alpha.3
+
+## 0.0.1-alpha.2
+
+### Patch Changes
+
+- 88600bc: Deployer fixes
+
 ## 0.0.1-alpha.1
 
 ### Patch Changes