+++ examples/agent/CHANGELOG.md
@@ -1,5 +1,52 @@
 # examples-agent
 
+## 1.0.1-alpha.50
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [c7abf8e]
+- Updated dependencies [9ade36e]
+  - @mastra/deployer-cloudflare@0.0.1-alpha.8
+  - @mastra/core@0.1.27-alpha.69
+
+## 1.0.1-alpha.49
+
+### Patch Changes
+
+- @mastra/deployer-cloudflare@0.0.1-alpha.7
+
+## 1.0.1-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/deployer-cloudflare@0.0.1-alpha.6
+
+## 1.0.1-alpha.47
+
+### Patch Changes
+
+- @mastra/deployer-cloudflare@0.0.1-alpha.5
+
+## 1.0.1-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [026ca5d]
+  - @mastra/deployer-cloudflare@0.0.1-alpha.4
+
+## 1.0.1-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+- Updated dependencies [a291824]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/deployer-cloudflare@0.0.1-alpha.3
+
 ## 1.0.1-alpha.44
 
 ### Patch Changes