+++ examples/integrations/CHANGELOG.md
@@ -1,5 +1,34 @@
 # mastra-examples-integrations
 
+## 1.0.1-alpha.39
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+  - @mastra/composio@0.0.1-alpha.35
+  - @mastra/firecrawl@1.0.4-alpha.57
+
+## 1.0.1-alpha.38
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/composio@0.0.1-alpha.34
+  - @mastra/firecrawl@1.0.4-alpha.56
+
+## 1.0.1-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/composio@0.0.1-alpha.33
+  - @mastra/firecrawl@1.0.4-alpha.55
+
 ## 1.0.1-alpha.36
 
 ### Patch Changes