+++ examples/memory/CHANGELOG.md
@@ -1,5 +1,31 @@
 # memory
 
+## 1.0.1-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+  - @mastra/memory@0.0.2-alpha.49
+
+## 1.0.1-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/memory@0.0.2-alpha.48
+
+## 1.0.1-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/memory@0.0.2-alpha.47
+
 ## 1.0.1-alpha.44
 
 ### Patch Changes