+++ examples/openapi-spec-writer/CHANGELOG.md
@@ -1,5 +1,41 @@
 # openapi-spec-writer
 
+## 0.1.1-alpha.53
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/loggers@0.0.1-alpha.0
+  - @mastra/core@0.1.27-alpha.69
+  - @mastra/rag@0.0.2-alpha.55
+  - @mastra/firecrawl@1.0.4-alpha.57
+  - @mastra/github@1.0.3-alpha.53
+  - @mastra/engine@0.0.5-alpha.63
+
+## 0.1.1-alpha.52
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/firecrawl@1.0.4-alpha.56
+  - @mastra/github@1.0.3-alpha.52
+  - @mastra/engine@0.0.5-alpha.62
+  - @mastra/rag@0.0.2-alpha.54
+
+## 0.1.1-alpha.51
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/firecrawl@1.0.4-alpha.55
+  - @mastra/github@1.0.3-alpha.51
+  - @mastra/engine@0.0.5-alpha.61
+  - @mastra/rag@0.0.2-alpha.53
+
 ## 0.1.1-alpha.50
 
 ### Patch Changes