+++ examples/travel-app/CHANGELOG.md
@@ -1,5 +1,34 @@
 # travel-app
 
+## 0.1.1-alpha.24
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+  - @mastra/engine@0.0.5-alpha.63
+  - @mastra/memory@0.0.2-alpha.49
+
+## 0.1.1-alpha.23
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/engine@0.0.5-alpha.62
+  - @mastra/memory@0.0.2-alpha.48
+
+## 0.1.1-alpha.22
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/engine@0.0.5-alpha.61
+  - @mastra/memory@0.0.2-alpha.47
+
 ## 0.1.1-alpha.21
 
 ### Patch Changes