+++ examples/workflow-ai-recruiter/CHANGELOG.md
@@ -1,5 +1,28 @@
 # workflow-ai-recruiter
 
+## 1.0.1-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+
+## 1.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 1.0.1-alpha.13
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 1.0.1-alpha.12
 
 ### Patch Changes