+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,51 @@
 # mastra
 
+## 0.1.57-alpha.107
+
+### Patch Changes
+
+- 1944807: Unified logger and major step in better logs
+- 6e18618: Generate tsconfig on mastra create
+- 9ade36e: Changed measure for evals, added endpoints, attached metrics to agent, added ui for evals in playground, and updated docs
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/deployer@0.0.1-alpha.7
+  - @mastra/core@0.1.27-alpha.69
+
+## 0.1.57-alpha.106
+
+### Patch Changes
+
+- Updated dependencies [291fe57]
+- Updated dependencies [1a41fbf]
+  - @mastra/deployer@0.0.1-alpha.6
+
+## 0.1.57-alpha.105
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/deployer@0.0.1-alpha.5
+
+## 0.1.57-alpha.104
+
+### Patch Changes
+
+- 7babd5c: CLI build and other
+- Updated dependencies [7babd5c]
+  - @mastra/deployer@0.0.1-alpha.4
+
+## 0.1.57-alpha.103
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+- Updated dependencies [a291824]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/deployer@0.0.1-alpha.3
+
 ## 0.1.57-alpha.102
 
 ### Patch Changes