+++ packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @mastra/core
 
+## 0.1.27-alpha.69
+
+### Patch Changes
+
+- 1944807: Unified logger and major step in better logs
+- 9ade36e: Changed measure for evals, added endpoints, attached metrics to agent, added ui for evals in playground, and updated docs
+
+## 0.1.27-alpha.68
+
+### Patch Changes
+
+- 0be7181: Fix forward version
+- 0be7181: Add perplexity models
+
+## 0.1.27-alpha.67
+
+### Patch Changes
+
+- c8ff2f5: Fixed passing CoreMessages to stream/generate where the role is not user. Previously all messages would be rewritten to have role: "user"
+
 ## 0.1.27-alpha.66
 
 ### Patch Changes