+++ packages/deployer/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @mastra/deployer
 
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- 1944807: Unified logger and major step in better logs
+- 9ade36e: Changed measure for evals, added endpoints, attached metrics to agent, added ui for evals in playground, and updated docs
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+
+## 0.0.1-alpha.6
+
+### Patch Changes
+
+- 291fe57: mastra openapi, swagger ui, dynamic servers
+- 1a41fbf: Fix playground workflow triggerData on execution
+
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- 7babd5c: CLI build and other
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- a291824: Deployer fixes
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes