+++ packages/rag/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @mastra/rag
 
+## 0.0.2-alpha.55
+
+### Patch Changes
+
+- 1944807: Unified logger and major step in better logs
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+
+## 0.0.2-alpha.54
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 0.0.2-alpha.53
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 0.0.2-alpha.52
 
 ### Patch Changes