+++ packages/tts/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @mastra/tts
 
+## 0.0.1-alpha.16
+
+### Patch Changes
+
+- Updated dependencies [1944807]
+- Updated dependencies [9ade36e]
+  - @mastra/core@0.1.27-alpha.69
+
+## 0.0.1-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 0.0.1-alpha.13
 
 ### Patch Changes